requirements.txt
fastapi
uvicorn
discord.py
requests